Initial source for this investigation
It turns out that I had misread the docs.
https://docs.automapper.org/en/v12.0.0/Configuration.html#profile-instances
The last paragraph here is what is relevant:

> Configuration inside a profile only applies to maps inside the profile. Configuration applied to the root configuration applies to all maps created.

So separating the global ignores or should-map logic
into their own profile actually ends up scoping them in such
a manner that they will never be used.  This is my key mistake.
I need to move that logic to the root of the configuration
outside any profile.